debezium/dbz#1097 Refactor MilvusTestResource location
