+++ debezium-platform-conductor/src/test/java/io/debezium/platform/environment/database/db/MilvusTestResource.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.platform.environment.database;
+package io.debezium.platform.environment.database.db;
 
 import java.util.Map;
